src/requirements.txt
# common
Flask==3.0.0

# auth
Flask-MySQLdb==2.0.0
PyJWT

# handling queue
pika